+++ src/utilities/index.ts
@@ -1,3 +1,4 @@
 export { checkPermission } from './checkPermission';
+export { getPermissionQuery } from './getPermissionQuery';
 export { hasPermission } from './hasPermission';